requirements.txt
boto==2.36.0
Collectfast==0.2.1
dj-database-url==0.3.0
Django==1.7.7
django-appconf==1.0.1
django-bower==5.0.2
django-braces==1.4.0
django-compressor==1.4
django-fanout==1.1.0
django-libsass==0.2
django-model-utils==2.2
django-storages-redux==1.2.2
djangorestframework==3.0.5
fanout==1.1.0
gunicorn==19.2.1
libsass==0.6.2
mock==1.0.1
model-mommy==1.2.4
psycopg2==2.6
pubcontrol==2.1.0
PyJWT==0.4.3
python-dateutil==2.4.0
python-decouple==2.3
pytz==2014.10
requests==2.5.1
responses==0.3.0
six==1.9.0
git+https://github.com/vintasoftware/django-naomi.git@fix-webbrowser-open
